Refactor: Extract duplicated syntax highlighting into shared module in webapp [S] (#259)
